+++ src/Codeception/Module/WPFilesystem.php
@@ -1104,7 +1104,7 @@ PHP;
 /*
 Theme Name: $name
 Author: wp-browser
-Description: $name 
+Description: $name
 Version: 1.0
 */
 CSS;